refactor: Provider folder renamed to Core
