+++ src/Stride.CommunityToolkit.Examples/Core/Example.cs
@@ -1,3 +1,3 @@
-namespace Stride.CommunityToolkit.Examples.Providers;
+namespace Stride.CommunityToolkit.Examples.Core;
 
 public record Example(string Id, string Title, Action Action);
