+++ src/Stride.CommunityToolkit.Examples/Core/ExampleProjectMeta.cs
@@ -1,4 +1,4 @@
-namespace Stride.CommunityToolkit.Examples.Providers;
+namespace Stride.CommunityToolkit.Examples.Core;
 
 public sealed record ExampleProjectMeta(
     string Id,
